ui/streamlit/requirements.txt
grpcio-tools==1.46.5
streamlit
